[ refactor ] seperate filter tag value with custom hook
